Change how position-based skipping works.
This gets rid of from() and until() and replaces them by set_front() and set_back(),
the latter of which is only in BidirectionalRange. The new functions modify the
current range instead of returning a new one.

This change makes the concepts simpler and more consistent. Now all modifiers
actually modify, instead of the weird difference between drop_front() and from().
It also gets rid of the weird until() that could return a different type, which
was probably completely useless in algorithms anyway. If it is really needed, it
can be emulated with a take_until range whose predicate works on the range and
tests a known position p against r.at_front().

I cannot think of a forward range where until() would have returned the same type
that isn't a bidirectional range anyway.